messaging fully refactored
