+++ backend/routes/internalOps.js
@@ -6,7 +6,7 @@
  *     /_____/  \____/  \__, /   \__, /  /_/     
  *                     /____/   /____/           
  *
- * @fileoverview This file contains the router definitions for internal API endpoints that manage user data.
+ * @fileoverview This file contains the router definitions for internal API endpoints that manage user data for Doggr.
  * These endpoints are used primarily for administrative tasks such as managing user accounts, interactions, and testing.
  * These routes are intended for use by system administrators and should ideally be protected by appropriate authentication mechanisms.
  * @author Carter VanHaren
